add set_ply_config.c
