+++ include/parse_ply.h
@@ -6,7 +6,7 @@
 /*   By: tyamagis <tyamagis@student.42tokyo.>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/16 14:40:51 by tyamagis          #+#    #+#             */
-/*   Updated: 2022/02/18 22:13:27 by tyamagis         ###   ########.fr       */
+/*   Updated: 2022/02/24 22:13:57 by tyamagis         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,10 +34,11 @@ typedef struct s_ply
 	t_face		*faces;
 }	t_ply;
 
-t_ply	*parse_ply(char *filename);
+void	*recreate_faces(t_ply *ply_info, t_face *old, int *read, int elm);
+void	*poly_to_tri(t_ply *ply, FILE *f_stream, int *read, int vtx);
 void	*set_vertexes(t_ply *ply, FILE *f_stream);
 void	*set_faces(t_ply *ply, FILE *f_stream);
-void	*poly_to_tri(t_ply *ply, FILE *f_stream, int *read, int vtx);
-void	*recreate_faces(t_ply *ply_info, t_face *old, int *read, int elm);
+void	check_ply_header(t_ply *p, FILE *f_stream);
+t_ply	*parse_ply(char *filename);
 
 #endif
