+++ loop_draw.c
@@ -6,7 +6,7 @@
 /*   By: tyamagis <tyamagis@student.42tokyo.>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/18 20:03:14 by tyamagis          #+#    #+#             */
-/*   Updated: 2022/02/24 21:55:25 by tyamagis         ###   ########.fr       */
+/*   Updated: 2022/02/24 22:18:43 by tyamagis         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
